Fix parse error against diffstat in git format-patch
The parser was not accepting diffstat from git format-patch files.
This patch fixes this issue and rewrites the git.diff file to contain
the full output of a git format-patch command to exercise the parser
(minus the git version, as this would trigger an assert error because of
remaining content).

Fixes #27